[#70] refactor: package 전략 변경
