+++ src/main/java/com/study/realworld/comment/domain/CommentRepository.java
@@ -1,0 +1,14 @@
+package com.study.realworld.comment.domain;
+
+import com.study.realworld.article.domain.Article;
+import java.util.List;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+
+    List<Comment> findAllByArticle(Article article);
+
+    Optional<Comment> findByIdAndArticle(Long id, Article article);
+
+}
