+++ src/main/java/com/study/realworld/global/security/JwtService.java
@@ -1,0 +1,14 @@
+package com.study.realworld.global.security;
+
+import com.study.realworld.user.domain.User;
+import java.util.Optional;
+import org.springframework.stereotype.Service;
+
+@Service
+public interface JwtService {
+
+    String createToken(User user);
+
+    Optional<User> getUser(String accessToken);
+
+}
