--- src/main/java/com/study/realworld/security/JwtAuthentication.java
@@ -1,28 +1,0 @@
-package com.study.realworld.security;
-
-import org.springframework.security.authentication.AbstractAuthenticationToken;
-
-public class JwtAuthentication extends AbstractAuthenticationToken {
-
-    private final Object principal;
-
-    private final String credentials;
-
-    public JwtAuthentication(Object principal, String credentials) {
-        super(null);
-        super.setAuthenticated(true);
-
-        this.principal = principal;
-        this.credentials = credentials;
-    }
-
-    @Override
-    public Object getCredentials() {
-        return credentials;
-    }
-
-    @Override
-    public Object getPrincipal() {
-        return principal;
-    }
-}
