+++ src/main/java/com/study/realworld/user/controller/UserController.java
@@ -2,7 +2,7 @@ package com.study.realworld.user.controller;
 
 import static com.study.realworld.user.controller.response.UserResponse.fromUserAndToken;
 
-import com.study.realworld.security.JwtService;
+import com.study.realworld.global.security.JwtService;
 import com.study.realworld.user.controller.request.UserJoinRequest;
 import com.study.realworld.user.controller.request.UserLoginRequest;
 import com.study.realworld.user.controller.request.UserUpdateRequest;
