--- src/test/java/com/study/realworld/article/comment/controller/response/CommentResponseTest.java
@@ -1,12 +1,0 @@
-package com.study.realworld.article.comment.controller.response;
-
-import org.junit.jupiter.api.Test;
-
-class CommentResponseTest {
-
-    @Test
-    void commentResponseTest() {
-        CommentResponse commentResponse = new CommentResponse();
-    }
-
-}
