--- src/test/java/com/study/realworld/article/comment/controller/response/CommentsResponseTest.java
@@ -1,12 +1,0 @@
-package com.study.realworld.article.comment.controller.response;
-
-import org.junit.jupiter.api.Test;
-
-class CommentsResponseTest {
-
-    @Test
-    void commentsResponseTest() {
-        CommentsResponse commentsResponse = new CommentsResponse();
-    }
-
-}
