+++ src/test/java/com/study/realworld/comment/controller/response/CommentResponseTest.java
@@ -1,0 +1,12 @@
+package com.study.realworld.comment.controller.response;
+
+import org.junit.jupiter.api.Test;
+
+class CommentResponseTest {
+
+    @Test
+    void commentResponseTest() {
+        CommentResponse commentResponse = new CommentResponse();
+    }
+
+}
