+++ src/test/java/com/study/realworld/comment/controller/response/CommentsResponseTest.java
@@ -1,0 +1,12 @@
+package com.study.realworld.comment.controller.response;
+
+import org.junit.jupiter.api.Test;
+
+class CommentsResponseTest {
+
+    @Test
+    void commentsResponseTest() {
+        CommentsResponse commentsResponse = new CommentsResponse();
+    }
+
+}
